requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
debugpy==1.5.0
decorator==5.1.0
entrypoints==0.3
Flask==2.0.2
Flask-PyMongo==2.3.0
greenlet==1.1.2
idna==3.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
nest-asyncio==1.5.1
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
passlib==1.7.4
pexpect==4.8.0
ptyprocess==0.7.0
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0
SQLAlchemy==1.4.26
urllib3==1.26.7
Werkzeug==2.0.2
